Use brew for parity Linux <> macOS
The differences in terrafrom_docs were due to go get getting the
latest version of the code. This can be fixed in 2 ways:
- specify version to go get
- use brew on linux

I initially did teh first but I realized that for each upgrade that
is done in Homebrew the version string would have to be updated.
Using brew ensures parity for versions with no effort. A downside
to this is the runtime taking longer, but I belive that is a good
tradeoff to make.