Remove conditional image fetch
This change will make it so that the module will always fetch both the
linux and windows AMI's. This is because, if you deploy out a cluster
that has a worker_group and then change the worker group, the change
ends up having an error like:

```
Error: Invalid count argument

  on .terraform/modules/cluster.eks/data.tf line 34, in data "aws_ami" "eks_worker_windows":
  34:   count = local.worker_has_windows_ami ? 1 : 0
```

Which can only be fixed by removing all the nodes from the worker_group,
applying the change, and then re-adding the worker groups.