Avoid using cloudinit at all unless the config specifies a MIM type
Prior to this change, node_groups userdata had to be specified as a
cloudinit config file.  However, Bottlerocket OS doesn't support
cloudinit and relies on bare TOML.

This change makes the use of cloudinit config rely on the specification
of a MIME type for the templated userdata file.  Otherwise, the userdata
provided falls back to the raw templated file content.

*N.B.*: This is backwards compatible, as the default `user_data`
specification includes a MIME type (`text/x-shellscript`).