Specify node_group user data via flat variables
Prior to this change, the user data templating variables were nested
under a top level `user_data` module variable, so that when specifying
one piece of it, the default behaviour (specifically the default MIME
type) changed.  There was no precedent in the module for this kind of
behaviour.

This change flattens the user data module variables, making them
independent of each other.  This requires a slightly odd behaviour
around the MIME type, which we have to specify as explicitly `""` (or
alternatively provide a slightly verbose additional module flag) in
order to prevent wrapping the user data in a CloudInit packet.