Make device name configurable for managed node groups
Prior to this change, the device name was hard coded to `/dev/xvda`,
while Bottlerocket has two volumes: a core OS volume that should be left
as defined by the AMI; and a user-space volume `/dev/xvdb` that _should_
be configurable (but wasn't).

This change allows the specification of the device name for the sole
configurable EBS volume in the managed node group launch template.  This
therefore supports the Bottlerocket usecase (as demonstrated in the
`managed_bottlerocket_node_group` example), as well as any other AMI
with alternative volume naming.