feat: Add AmazonEKSServicePolicy to clusters for compatibility
https://docs.aws.amazon.com/eks/latest/userguide/security-iam-awsmanpol.html

You can attach AmazonEKSServicePolicy to your IAM entities. Clusters that
were created before April 16, 2020, required you to create an IAM role
and attach this policy to it. Clusters that were created on or after
April 16, 2020, don't require you to create a role and don't require
you to assign this policy.