Use MIME format for user data
When enable_bootstrap_user_data: true, Amazon EC2 user data in launch
templates that are used with managed node groups must be in the MIME
multi-part archive format. This is because user data is merged with
Amazon EKS user data required for nodes to join the cluster.

It looks like AWS API accepts user data not in MIME multi-part format
when creating an EKS node group with platform linux (custom AMI types),
but checks on update and errors out:

Ec2LaunchTemplateInvalidConfiguration: User data was not in the MIME
multipart format.

With that change, pre_/post_ bootstrap_user_data will be wrapped with
MIME tags as required, if using enable_bootstrap_user_data, or will be
skipped otherwise. This also provides a fix for the case when
pre_bootstrap_user_data is still injected into user data, when
the enable_bootstrap_user_data flag is unset.

Signed-off-by: Bogdan Dobrelya <bogdando@yahoo.com>