Add option to set instance_type on the launch template
This allows setting the instance type on the launch template rather than
the EKS Managed node group. This is useful because it is not possible to
change the instance types set on a managed node groups after creation.
The suggested way to work around this is to simply define the instance
type in the launch template.

https://github.com/aws/containers-roadmap/issues/746#issuecomment-675099508