fix: optionalize cluster endpoint and cert data in templates
The parameters should not be rendered in the template if they are not passed properly. This is the case, when the cluster is configured for calico or another CNI plugin. The values are optional in terraform and should only be rendered when passed, else it creates an invalid bash command with possible unintended side effects.

The bootstrap script still works without the values due to an internal fallback to a describe cluster call.

fixes #2787