fix: more robust and cleaner compute_config
When not using eks-auto mode, the compute_config for_each condition checks the length of var.cluster_compute_config, which works fine if the variable is not specified. However, if a downstream consumer wants to create a module that toggles eks-auto, then cluster_compute_config must be defined. This change makes the module more robust and ensures it can support both eks-auto and standard EKS configurations.